PG metrics DB: allow per-table dropping of old data partitions
It was reported that for a very large amount of monitored DB-s one can
currently easilt get errors when dropping old data partitions as
Postgres is running out of max_locks_per_transaction. The solution for
metric-time and metric-dbname-time schemas is to drop tables 1-by-1.
Needs investigation if this is a problem also with TimescaleDB schema.