Add "prerequisite extensions" support for metric definitions
To reduce regular errors in server logs of monitored DBs. That is - we
don't even try to query the DB if we know that some prerequisite is missing
and it would error anyways. In the pgwatch2 logs the errors got reduced already
to 1x per 10min, but increase that to 1h also for this case.

For now only for pg_stat_statements as that's by far the most common one
in use.