[-] fix warning about yaml.load() is deprecated, closes #738 (#739)
This is a fix for deprecated use of the yaml.load function. Apparently yaml.load carries some security concerns with how it de-serializes data from the yaml file, (arbitrary python code could be executed if embedded in the yaml file). As a result the library maintainers want users to be more explicit about what loader they are using to load the yaml.

The current usage of yaml.load in pgwatch2 is equivalent to the yaml.full_load function used in this PR. A safer choice might be the yaml.safe_load function but I would be concerned about regression issues.